Revert "wip-time handleNewBooking refactory"
This reverts commit 6babb31dbdf2991747a6b62cd34d4443814d5144.
