--- packages/features/bookings/lib/handleBookingSimpleChange.ts
@@ -1,21 +1,0 @@
-// Placeholder for handleBookingSimpleChange.ts
-// This file will eventually use the onBookingCreationInDb function
-// after performing its specific simple change logic and potentially re-fetching booking/event data.
-// import { onBookingCreationInDb } from "./onBookingCreationInDb";
-
-export async function handleBookingSimpleChange() {
-  // Define parameters for handling a simple change, e.g., bookingId, changeDetails
-  // This function would fetch the booking, apply changes, then potentially call
-  // onBookingCreationInDb if similar post-update logic is needed.
-  console.log("handleBookingSimpleChange called - to be implemented");
-
-  // Example structure (very-high level):
-  // 1. Fetch existing booking by ID
-  // 2. Apply simple changes to booking data / evt data
-  // 3. Potentially call prisma.booking.update for simple field changes not covered by onBookingCreationInDb's scope
-  // 4. Prepare all parameters for onBookingCreationInDb
-  // 5. const { videoCallUrl, paymentDetails } = await onBookingCreationInDb({ ...params });
-  // 6. Return result
-
-  return { message: "Function not yet implemented." };
-}
